Fix tests for ObjectFileGenerator and ObjectTemplate
